+++ apps/prairielearn/src/ee/auth/lti13/lti13Auth.ts
@@ -1,22 +1,26 @@
+import * as crypto from 'crypto';
+import { URL } from 'url';
+import { callbackify } from 'util';
+
 import { Router, type Request, Response, NextFunction } from 'express';
-import asyncHandler = require('express-async-handler');
+import asyncHandler from 'express-async-handler';
+import _ from 'lodash';
 import { Issuer, Strategy, type TokenSet } from 'openid-client';
 import * as passport from 'passport';
 import { z } from 'zod';
-import { callbackify } from 'util';
-import * as crypto from 'crypto';
-import { URL } from 'url';
-import { Lti13ClaimType, Lti13ClaimSchema, Lti13Claim } from '../../lib/lti13';
 
-import { loadSqlEquiv, queryAsync } from '@prairielearn/postgres';
-import * as error from '@prairielearn/error';
 import { cache } from '@prairielearn/cache';
-import * as authnLib from '../../../lib/authn';
-import { selectLti13Instance } from '../../models/lti13Instance';
-import { Lti13Test } from './lti13Auth.html';
-import { getCanonicalHost } from '../../../lib/url';
+import * as error from '@prairielearn/error';
+import { loadSqlEquiv, queryAsync } from '@prairielearn/postgres';
 
-const sql = loadSqlEquiv(__filename);
+import * as authnLib from '../../../lib/authn.js';
+import { getCanonicalHost } from '../../../lib/url.js';
+import { Lti13ClaimType, Lti13ClaimSchema, Lti13Claim } from '../../lib/lti13.js';
+import { selectLti13Instance } from '../../models/lti13Instance.js';
+
+import { Lti13Test } from './lti13Auth.html.js';
+
+const sql = loadSqlEquiv(import.meta.url);
 const router = Router({ mergeParams: true });
 
 const StateTest = '-StateTest';