+++ apps/prairielearn/src/ee/lib/selectAndAuthz.ts
@@ -1,9 +1,11 @@
 import { z } from 'zod';
-import { loadSqlEquiv, queryOptionalRow } from '@prairielearn/postgres';
-import { AdministratorSchema, type Institution, InstitutionSchema } from '../../lib/db-types';
-import { HttpStatusError } from '@prairielearn/error';
 
-const sql = loadSqlEquiv(__filename);
+import { HttpStatusError } from '@prairielearn/error';
+import { loadSqlEquiv, queryOptionalRow } from '@prairielearn/postgres';
+
+import { AdministratorSchema, type Institution, InstitutionSchema } from '../../lib/db-types.js';
+
+const sql = loadSqlEquiv(import.meta.url);
 
 export async function selectAndAuthzInstitutionAsAdmin({
   institution_id,