+++ apps/prairielearn/src/models/questions.ts
@@ -1,13 +1,15 @@
+import { AnsiUp } from 'ansi_up';
+import { z } from 'zod';
+
 import * as sqldb from '@prairielearn/postgres';
-import AnsiUp from 'ansi_up';
+
 import {
   TopicSchema,
   SharingSetSchema,
   AssessmentsFormatForQuestionSchema,
   TagSchema,
-} from '../lib/db-types';
-import { z } from 'zod';
-import { idsEqual } from '../lib/id';
+} from '../lib/db-types.js';
+import { idsEqual } from '../lib/id.js';
 
 const QuestionsPageDataSchema = z.object({
   id: z.string(),
@@ -33,7 +35,7 @@ export interface QuestionsPageDataAnsified extends QuestionsPageData {
 }
 
 const ansiUp = new AnsiUp();
-const sql = sqldb.loadSqlEquiv(__filename);
+const sql = sqldb.loadSqlEquiv(import.meta.url);
 
 export async function selectQuestionsForCourse(
   course_id: string | number,