+++ apps/prairielearn/src/pages/editError/editError.js
@@ -1,10 +1,12 @@
 // @ts-check
-const asyncHandler = require('express-async-handler');
 import { Router } from 'express';
+import asyncHandler from 'express-async-handler';
+
 import { HttpStatusError } from '@prairielearn/error';
-import { getJobSequenceWithFormattedOutput } from '../../lib/server-jobs';
-import * as syncHelpers from '../shared/syncHelpers';
 import { logger } from '@prairielearn/logger';
+
+import { getJobSequenceWithFormattedOutput } from '../../lib/server-jobs.js';
+import * as syncHelpers from '../shared/syncHelpers.js';
 
 const router = Router();
 
@@ -44,7 +46,7 @@ router.get(
     }
 
     res.locals.job_sequence = job_sequence;
-    res.render(__filename.replace(/\.js$/, '.ejs'), res.locals);
+    res.render(import.meta.filename.replace(/\.js$/, '.ejs'), res.locals);
   }),
 );