+++ apps/prairielearn/src/pages/publicQuestionPreview/publicQuestionPreview.ts
@@ -1,20 +1,22 @@
 import { Router } from 'express';
-import * as error from '@prairielearn/error';
+import asyncHandler from 'express-async-handler';
 import { z } from 'zod';
-import asyncHandler = require('express-async-handler');
 
-import { selectQuestionById } from '../../models/question';
-import { selectCourseById } from '../../models/course';
-import { processSubmission } from '../../lib/question-submission';
-import { IdSchema, UserSchema } from '../../lib/db-types';
-import { logPageView } from '../../middlewares/logPageView';
+import * as error from '@prairielearn/error';
+
+import { setQuestionCopyTargets } from '../../lib/copy-question.js';
+import { IdSchema, UserSchema } from '../../lib/db-types.js';
 import {
   getAndRenderVariant,
   renderPanelsForSubmission,
   setRendererHeader,
-} from '../../lib/question-render';
-import { PublicQuestionPreview } from './publicQuestionPreview.html';
-import { setQuestionCopyTargets } from '../../lib/copy-question';
+} from '../../lib/question-render.js';
+import { processSubmission } from '../../lib/question-submission.js';
+import { logPageView } from '../../middlewares/logPageView.js';
+import { selectCourseById } from '../../models/course.js';
+import { selectQuestionById } from '../../models/question.js';
+
+import { PublicQuestionPreview } from './publicQuestionPreview.html.js';
 
 const router = Router({ mergeParams: true });