+++ apps/prairielearn/src/pages/studentAssessment/studentAssessment.js
@@ -1,13 +1,12 @@
 // @ts-check
-const asyncHandler = require('express-async-handler');
 import * as express from 'express';
+import asyncHandler from 'express-async-handler';
 
 import * as error from '@prairielearn/error';
-import { loadSqlEquiv, queryAsync } from '@prairielearn/postgres';
 import { flash } from '@prairielearn/flash';
+import { loadSqlEquiv, queryAsync } from '@prairielearn/postgres';
 
-import { checkPasswordOrRedirect } from '../../middlewares/studentAssessmentAccess';
-import { makeAssessmentInstance } from '../../lib/assessment';
+import { makeAssessmentInstance } from '../../lib/assessment.js';
 import {
   joinGroup,
   createGroup,
@@ -18,11 +17,12 @@ import {
   leaveGroup,
   GroupOperationError,
   canUserAssignGroupRoles,
-} from '../../lib/groups';
-import { getClientFingerprintId } from '../../middlewares/clientFingerprint';
+} from '../../lib/groups.js';
+import { getClientFingerprintId } from '../../middlewares/clientFingerprint.js';
+import { checkPasswordOrRedirect } from '../../middlewares/studentAssessmentAccess.js';
 
 const router = express.Router();
-const sql = loadSqlEquiv(__filename);
+const sql = loadSqlEquiv(import.meta.url);
 
 router.get(
   '/',
@@ -75,7 +75,7 @@ router.get(
           }
         }
       }
-      res.render(__filename.replace(/\.js$/, '.ejs'), res.locals);
+      res.render(import.meta.filename.replace(/\.js$/, '.ejs'), res.locals);
     } else {
       const result = await queryAsync(sql.select_single_assessment_instance, params);
       if (result.rowCount === 0) {
@@ -112,7 +112,7 @@ router.get(
               );
             }
           }
-          res.render(__filename.replace(/\.js$/, '.ejs'), res.locals);
+          res.render(import.meta.filename.replace(/\.js$/, '.ejs'), res.locals);
         } else if (res.locals.assessment.type === 'Homework') {
           const time_limit_min = null;
           const client_fingerprint_id = await getClientFingerprintId(req, res);
@@ -128,7 +128,7 @@ router.get(
           );
           res.redirect(res.locals.urlPrefix + '/assessment_instance/' + assessment_instance_id);
         } else {
-          res.render(__filename.replace(/\.js$/, '.ejs'), res.locals);
+          res.render(import.meta.filename.replace(/\.js$/, '.ejs'), res.locals);
         }
       } else {
         res.redirect(res.locals.urlPrefix + '/assessment_instance/' + result.rows[0].id);