+++ apps/prairielearn/src/tests/accessibility/index.test.ts
@@ -1,22 +1,18 @@
-// The OpenTelemetry instrumentation for Express breaks our ability to inspect
-// the Express routes. We need to disable it before loading the server.
-import { disableInstrumentations } from '@prairielearn/opentelemetry';
-disableInstrumentations();
-
-import { test } from 'mocha';
-import axe = require('axe-core');
-import { JSDOM } from 'jsdom';
-import fetch from 'node-fetch';
 import { A11yError } from '@sa11y/format';
-import expressListEndpoints = require('express-list-endpoints');
+import axe from 'axe-core';
+import expressListEndpoints from 'express-list-endpoints';
+import { JSDOM } from 'jsdom';
+import { test } from 'mocha';
+import fetch from 'node-fetch';
+
 import * as sqldb from '@prairielearn/postgres';
 
-import * as server from '../../server';
-import * as news_items from '../../news_items';
-import { config } from '../../lib/config';
-import * as helperServer from '../helperServer';
-import { features } from '../../lib/features/index';
-import { EXAMPLE_COURSE_PATH } from '../../lib/paths';
+import { config } from '../../lib/config.js';
+import { features } from '../../lib/features/index.js';
+import { EXAMPLE_COURSE_PATH } from '../../lib/paths.js';
+import * as news_items from '../../news_items/index.js';
+import * as server from '../../server.js';
+import * as helperServer from '../helperServer.js';
 
 const SITE_URL = 'http://localhost:' + config.serverPort;
 
@@ -256,7 +252,7 @@ describe('accessibility', () => {
       errorIfLockNotAcquired: true,
     });
 
-    const app = server.initExpress();
+    const app = await server.initExpress();
     endpoints = expressListEndpoints(app);
     endpoints.sort((a, b) => a.path.localeCompare(b.path));