+++ apps/prairielearn/src/tests/groupGenerateAndDelete.test.ts
@@ -1,13 +1,15 @@
 import { assert } from 'chai';
-import * as sqldb from '@prairielearn/postgres';
 import { step } from 'mocha-steps';
 
-import * as helperServer from './helperServer';
-import { deleteAllGroups } from '../lib/groups';
-import * as groupUpdate from '../lib/group-update';
-import { TEST_COURSE_PATH } from '../lib/paths';
+import * as sqldb from '@prairielearn/postgres';
 
-const sql = sqldb.loadSqlEquiv(__filename);
+import * as groupUpdate from '../lib/group-update.js';
+import { deleteAllGroups } from '../lib/groups.js';
+import { TEST_COURSE_PATH } from '../lib/paths.js';
+
+import * as helperServer from './helperServer.js';
+
+const sql = sqldb.loadSqlEquiv(import.meta.url);
 const locals: Record<string, any> = {};
 
 describe('test auto group and delete groups', function () {