+++ packages/compiled-assets/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @prairielearn/compiled-assets
 
+## 3.0.1
+
+### Patch Changes
+
+- 901fce8: Upgrade all JavaScript dependencies
+- Updated dependencies [901fce8]
+  - @prairielearn/html@4.0.1
+
+## 3.0.0
+
+### Major Changes
+
+- 4f30b7e: Publish as native ESM
+
+### Patch Changes
+
+- Updated dependencies [4f30b7e]
+  - @prairielearn/html@4.0.0
+
 ## 2.0.17
 
 ### Patch Changes