+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @prairielearn/postgres-tools
 
+## 2.0.1
+
+### Patch Changes
+
+- 901fce8: Upgrade all JavaScript dependencies
+- Updated dependencies [901fce8]
+  - @prairielearn/postgres@2.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- 4f30b7e: Publish as native ESM
+
+### Patch Changes
+
+- Updated dependencies [4f30b7e]
+  - @prairielearn/postgres@2.0.0
+
 ## 1.1.17
 
 ### Patch Changes